Finalizado geração de relatorio de veiculos mais publicados

+++ src/dao/NivelAcessoJuridicoTelaSoftare.java
@@ -4,9 +4,9 @@ import model.DatabaseUtils;
 
 public class NivelAcessoJuridicoTelaSoftare extends DatabaseUtils {
 	private String nome_tabela = "nivel_acesso_juridico_tela_software";
-	private long idNivelAcessoJuridico;
-	private long idPermissaoJuridico;
-	private long idTelaSoftware;
+	private Integer idNivelAcessoJuridico;
+	private Integer idPermissaoJuridico;
+	private Integer idTelaSoftware;
 	
 	public String getNome_tabela() {
 		return nome_tabela;
@@ -17,19 +17,19 @@ public class NivelAcessoJuridicoTelaSoftare extends DatabaseUtils {
 	public long getIdNivelAcessoJuridico() {
 		return idNivelAcessoJuridico;
 	}
-	public void setIdNivelAcessoJuridico(long idNivelAcessoJuridico) {
+	public void setIdNivelAcessoJuridico(Integer idNivelAcessoJuridico) {
 		this.idNivelAcessoJuridico = idNivelAcessoJuridico;
 	}
 	public long getIdPermissaoJuridico() {
 		return idPermissaoJuridico;
 	}
-	public void setIdPermissaoJuridico(long idPermissaoJuridico) {
+	public void setIdPermissaoJuridico(Integer idPermissaoJuridico) {
 		this.idPermissaoJuridico = idPermissaoJuridico;
 	}
 	public long getIdTelaSoftware() {
 		return idTelaSoftware;
 	}
-	public void setIdTelaSoftware(long idTelaSoftware) {
+	public void setIdTelaSoftware(Integer idTelaSoftware) {
 		this.idTelaSoftware = idTelaSoftware;
 	}
